simplify proxy
